Rename Signatory to SignatoryContact, small reformatting
